requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
asyncpg>=0.29.0
pydantic>=2.7.0
numpy>=1.26.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
supabase-py>=1.0.0
